Updated files to use config
