+++ fluentai/constants/config.py
@@ -4,4 +4,4 @@ import yaml
 
 config_path = os.path.join(os.path.dirname(__file__), "..", "..", "config.yaml")
 with open(config_path, "r", encoding="utf-8") as f:
-    config = yaml.full_load(f)
+    config: dict = yaml.full_load(f)
