+++ fluentai/mnemonic/phonetic/phonetic.py
@@ -3,6 +3,7 @@ import numpy as np
 import pandas as pd
 from huggingface_hub import hf_hub_download
 
+from fluentai.constants.config import config
 from fluentai.mnemonic.phonetic.g2p import g2p
 from fluentai.mnemonic.phonetic.ipa2vec import panphon_vec, soundvec
 from fluentai.mnemonic.phonetic.utils import convert_to_matrix, load_cache, pad_vectors
@@ -19,8 +20,8 @@ def word2ipa(
         # First try lookup in the .tsv file
         eng_ipa = pd.read_csv(
             hf_hub_download(
-                repo_id="StephanAkkerman/english-words-IPA",
-                filename="en_US.csv",
+                repo_id=config.get("PHONETIC_SIM").get("IPA_REPO"),
+                filename=config.get("PHONETIC_SIM").get("IPA_FILE"),
                 cache_dir="datasets",
                 repo_type="dataset",
             )
